path changing: task-timer.model moved under own folder
